Rename global_level_events.hpp -> global.hpp
