+++ src/common/global.hpp
