+++ src/game_logic/dynamic_geometry_system.cpp
@@ -17,7 +17,7 @@
 #include "dynamic_geometry_system.hpp"
 
 #include "common/game_service_provider.hpp"
-#include "common/global_level_events.hpp"
+#include "common/global.hpp"
 #include "data/map.hpp"
 #include "data/sound_ids.hpp"
 #include "engine/base_components.hpp"
