+++ src/game_logic/interaction/elevator.cpp
@@ -17,7 +17,7 @@
 #include "elevator.hpp"
 
 #include "common/game_service_provider.hpp"
-#include "common/global_level_events.hpp"
+#include "common/global.hpp"
 #include "engine/collision_checker.hpp"
 #include "engine/entity_tools.hpp"
 #include "engine/physical_components.hpp"
