+++ src/game_logic/player/damage_system.cpp
@@ -16,7 +16,7 @@
 
 #include "damage_system.hpp"
 
-#include "common/global_level_events.hpp"
+#include "common/global.hpp"
 #include "data/player_model.hpp"
 #include "engine/base_components.hpp"
 #include "engine/physical_components.hpp"
